+++ ng_trajectory/criterions/profile/main.py
@@ -44,6 +44,7 @@ P.createAdd("_lr", 0.139, float, "Distance from center of mass to the rear axle 
 P.createAdd("reference", None, str, "Name of the file to load (x, y, t) reference path that cannot be close.", "init")
 P.createAdd("reference_dist", 1.0, float, "Minimum allowed distance from the reference at given time [m].", "init")
 P.createAdd("reference_rotate", 0, int, "Number of points to rotate the reference trajectory.", "init")
+P.createAdd("reference_laptime", 0, float, "Lap time of the given reference. 0 = estimated from data", "init")
 P.createAdd("save_solution_csv", "$", str, "When non-empty, save final trajectory to this file as CSV. Use '$' to use log name instead.", "init")
 P.createAdd("plot", False, bool, "Whether a graphical representation should be created.", "init (viz.)")
 P.createAdd("plot_reference", False, bool, "Whether the reference trajectory should be plotted.", "init (viz.)")
@@ -79,8 +80,18 @@ def init(**kwargs) -> None:
 
     if P.getValue("reference") is not None:
         REFERENCE = numpy.load(P.getValue("reference"))
-        REFERENCE = numpy.hstack((numpy.roll(REFERENCE[:, :2], -P.getValue("reference_rotate"), axis=0), REFERENCE[:, 2:]))
-        print ("Loaded reference with '%d' points." % len(REFERENCE), file = kwargs.get("logfile", sys.stdout))
+        #REFERENCE = numpy.hstack((numpy.roll(REFERENCE[:, :2], -P.getValue("reference_rotate"), axis=0), REFERENCE[:, 2:]))
+        # TODO: Lap time should be given, not estimated like this.
+        lap_time = P.getValue("reference_laptime")
+
+        if lap_time == 0.0:
+            # Lap time estimate
+            lap_time = REFERENCE[-1, 2] + numpy.mean([REFERENCE[-1, 2]-REFERENCE[-2, 2], REFERENCE[1, 2]-REFERENCE[0, 2]])
+
+        REFERENCE = numpy.roll(REFERENCE, -P.getValue("reference_rotate"), axis=0)
+        REFERENCE[:, 2] = REFERENCE[:, 2] - REFERENCE[0, 2]
+        REFERENCE[REFERENCE[:, 2] < 0, 2] += lap_time
+        print ("Loaded reference with '%d' points, lap time %fs." % (len(REFERENCE), lap_time), file = kwargs.get("logfile", sys.stdout))
     else:
         REFERENCE = None