+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 - Expose all parameters in plotter for `figureSave`, `trackPlot`, `bordersPlot` and `indicesPlot`.
 
+### Fixed
+- Bound version of `bayesian-optimization<=1.4.0` to resolve import errors.
+
 ## 1.9.4 - 2023-06-29
 ### Added
 - Criterions